backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
mysql-connector-python==9.0.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.2.0
python-dotenv==1.0.1
sqlalchemy==2.0.35
